backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1  # Revert to 2.5.1
SQLAlchemy==1.4.50      # Pin to 1.4.50 for compatibility
Flask-Cors==3.0.10
Flask-Login==0.5.0
psycopg2-binary==2.9.1
werkzeug==2.0.1
python-dotenv==0.19.0
numpy==1.21.2
Flask-Migrate==3.1.0